feat: implement scalar 64bit barret mul, prepare test to verify asm implementation

+++ tests/barret-mul/barret-mul.h
@@ -3,6 +3,6 @@
 #include <stdint.h>
 #include <stdio.h>
 
-extern uint32_t barret_mul(uint32_t a, uint32_t b, uint32_t n, uint32_t k, uint64_t C);
+extern uint64_t barret_mul(uint64_t a, uint64_t b, uint64_t n, uint64_t k, uint64_t C);
 __attribute__((optimize("O0"))) void empty_dummy() {}
 #endif
